Moves Range to tiledb::type namespace
