+++ test/src/unit-Reader.cc
@@ -39,6 +39,7 @@
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
 #include "tiledb/sm/misc/types.h"
 #include "tiledb/sm/query/reader.h"
+#include "tiledb/type/range/range.h"
 
 #ifdef _WIN32
 #include "tiledb/sm/filesystem/win.h"
@@ -50,6 +51,7 @@
 #include <iostream>
 
 using namespace tiledb::sm;
+using namespace tiledb::type;
 using namespace tiledb::test;
 
 /* ********************************* */
