+++ tiledb/sm/array_schema/attribute.h
@@ -39,6 +39,11 @@
 #include "tiledb/sm/misc/types.h"
 
 using namespace tiledb::common;
+using namespace tiledb::type;
+
+namespace tiledb::type {
+class Range;
+}
 
 namespace tiledb {
 namespace sm {
