+++ tiledb/sm/fragment/fragment_metadata.cc
@@ -51,6 +51,7 @@
 #include "tiledb/sm/tile/generic_tile_io.h"
 #include "tiledb/sm/tile/tile.h"
 #include "tiledb/sm/tile/tile_metadata_generator.h"
+#include "tiledb/type/range/range.h"
 
 #include <cassert>
 #include <iostream>
@@ -58,6 +59,7 @@
 #include <string>
 
 using namespace tiledb::common;
+using namespace tiledb::type;
 
 namespace tiledb {
 namespace sm {
