+++ tiledb/sm/subarray/range_subset.cc
@@ -35,6 +35,7 @@
 #include <iostream>
 
 using namespace tiledb::common;
+using namespace tiledb::type;
 
 namespace tiledb::sm {
