+++ tiledb/sm/subarray/subarray.h
@@ -56,6 +56,11 @@
 #include <vector>
 
 using namespace tiledb::common;
+using namespace tiledb::type;
+
+namespace tiledb::type {
+class Range;
+}
 
 namespace tiledb {
 namespace sm {
