+++ tiledb/type/range/CMakeLists.txt
@@ -1,5 +1,6 @@
 #
-# tiledb/sm/subarray/CMakeLists.txt
+# tiledb/type/range/CMakeLists.txt
+#
 #
 # The MIT License
 #
@@ -26,22 +27,11 @@
 
 include(common NO_POLICY_SCOPE)
 
-#
-# `range_subset` object library
-#
-add_library(range_subset OBJECT range_subset.cc)
-target_link_libraries(range_subset PUBLIC baseline $<TARGET_OBJECTS:baseline>)
-target_link_libraries(range_subset PUBLIC constants $<TARGET_OBJECTS:constants>)
-target_link_libraries(range_subset PUBLIC thread_pool $<TARGET_OBJECTS:thread_pool>)
-target_link_libraries(range_subset PUBLIC range $<TARGET_OBJECTS:range>)
+# Range library
+add_library(range OBJECT range.h)
+target_link_libraries(range PUBLIC baseline $<TARGET_OBJECTS:baseline>)
 
-#
 # Test-compile of object library ensures link-completeness
-#
-add_executable(compile_range_subset EXCLUDE_FROM_ALL)
-target_link_libraries(compile_range_subset PRIVATE range_subset)
-target_sources(compile_range_subset PRIVATE test/compile_range_subset_main.cc)
-
-if (TILEDB_TESTS)
-    add_subdirectory(test)
-endif()
+add_executable(compile_range EXCLUDE_FROM_ALL)
+target_sources(compile_range PRIVATE test/compile_range_main.cc)
+target_link_libraries(compile_range PRIVATE range)
